MVP Development
For this week I made a prototype for the rock, paper, scissors game I was working on. It unfortunately didn’t go as expected in any sense. I have looked at the code and all of the function’s multiple times, but I can’t seem to figure out what was wrong. If the game was working you are supposed to click on the symbols for rock, paper, or scissors. The AI you are going up against will choose one of those three options and try to beat you in the game. Depending on the result the text box will change to resemble said result. Not only is the result text box not showing but the set of images that represents the AI’s choice is also not showing. I would like to say that I’ve been looking at this problem for about 2 hours now, despite that everything looks fine, and it could be something I did like choose something specific in the project settings or it could be because I used the mobile template when I should’ve done the normal 2D template.
My plan for next week is that I’m going to take another look at it and decide on whether to re-do the whole thing. It may not work but it has helped me a number of times before. If I’m able to get it working, I will start making custom sprites for the buttons instead of using placeholders. I will also try to make it, so the orientation is in portrait and not landscape. If I get everything done easily I’ll even start trying to draw the monkeys and get them implemented into the game.